+++ get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: cfamilar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 16:30:49 by cfamilar          #+#    #+#             */
-/*   Updated: 2022/09/01 16:32:10 by cfamilar         ###   ########.fr       */
+/*   Updated: 2022/09/07 18:12:42 by cfamilar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,16 +15,17 @@
 
 # include <unistd.h>
 # include <stdlib.h>
+# include <stdio.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
 # endif
 
 char	*ft_strchr(char *s, int c);
-char	*ft_substr(char *s, unsigned int start, size_t len);
+char	*ft_substr(char *s, int start, int len);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strdup(char *s1);
 char	*get_next_line(int fd);
-size_t	ft_strlen(char *s);
+int		ft_strlen(char *s);
 
 #endif
